+++ libs/ymir-core/include/ymir/hw/vdp/vdp.hpp
@@ -867,6 +867,7 @@ private:
             lastCharIndex = 0xFFFFFFFF;
 
             bitmapData.fill(0);
+            bitmapDataAddress = 0xFFFFFFFF;
 
             lastVCellScroll = 0xFFFFFFFF;
         }